fix: fix some bugs with cluster
