+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start:prod": "tsc && node dist/index.js",
     "start:dev": "nodemon",
-    "start:multi": "ts-node ./src/multi/multi.ts",
+    "start:multi": "ts-node src/multi.ts",
     "test": "mocha --require @babel/register src/test/test.js --timeout 10000"
   },
   "repository": {
