+++ src/multi.ts
@@ -1,9 +1,8 @@
-import http from 'http';
 import cluster from 'cluster';
 import os from 'os';
 import 'dotenv/config';
 
-import { server } from '../index';
+import { server } from './server';
 
 const PORT = process.env.PORT || 8000;
 const numCPUs = os.cpus().length;
@@ -19,7 +18,9 @@ if (cluster.isPrimary) {
     console.log(`worker ${worker.process.pid} died`);
   });
 } else {
-  server.listen(PORT);
+  server.listen(PORT, () => {
+    console.log(`Server running on port ${PORT}`);
+  });
 
   console.log(`Worker ${process.pid} started`);
 }
